Rename useUpdateAtom to useSetAtom
Move useAtomValue and useSetAtom to core
