+++ tests/useSetAtom.test.tsx
@@ -1,9 +1,8 @@
 import { StrictMode, useEffect, useRef } from 'react'
 import type { PropsWithChildren } from 'react'
 import { fireEvent, render, waitFor } from '@testing-library/react'
-import { atom, useAtom } from 'jotai'
-import { useUpdateAtom } from 'jotai/utils'
-import { getTestProvider } from '../testUtils'
+import { atom, useAtom, useSetAtom } from 'jotai'
+import { getTestProvider } from './testUtils'
 
 const Provider = getTestProvider()
 
@@ -15,7 +14,7 @@ const useCommitCount = () => {
   return commitCountRef.current
 }
 
-it('useUpdateAtom does not trigger rerender in component', async () => {
+it('useSetAtom does not trigger rerender in component', async () => {
   const countAtom = atom(0)
 
   const Displayer = () => {
@@ -29,7 +28,7 @@ it('useUpdateAtom does not trigger rerender in component', async () => {
   }
 
   const Updater = () => {
-    const setCount = useUpdateAtom(countAtom)
+    const setCount = useSetAtom(countAtom)
     const commits = useCommitCount()
     return (
       <>
@@ -79,7 +78,7 @@ it('useUpdateAtom does not trigger rerender in component', async () => {
   })
 })
 
-it('useUpdateAtom with scope', async () => {
+it('useSetAtom with scope', async () => {
   const scope = Symbol()
   const countAtom = atom(0)
 
@@ -89,7 +88,7 @@ it('useUpdateAtom with scope', async () => {
   }
 
   const Updater = () => {
-    const setCount = useUpdateAtom(countAtom, scope)
+    const setCount = useSetAtom(countAtom, scope)
     return (
       <button onClick={() => setCount((value) => value + 1)}>increment</button>
     )
@@ -121,7 +120,7 @@ it('useUpdateAtom with scope', async () => {
   })
 })
 
-it('useUpdateAtom with write without an argument', async () => {
+it('useSetAtom with write without an argument', async () => {
   const countAtom = atom(0)
   const incrementCountAtom = atom(null, (get, set) =>
     set(countAtom, get(countAtom) + 1)
@@ -137,7 +136,7 @@ it('useUpdateAtom with write without an argument', async () => {
   }
 
   const Updater = () => {
-    const setCount = useUpdateAtom(incrementCountAtom)
+    const setCount = useSetAtom(incrementCountAtom)
     return <Button cb={setCount}>increment</Button>
   }
