+++ tests/utils/atomFamily.test.tsx
@@ -2,8 +2,7 @@ import { StrictMode, Suspense, useEffect, useRef, useState } from 'react'
 import { fireEvent, render, waitFor } from '@testing-library/react'
 import ReactDOM from 'react-dom'
 import { atom, useAtom } from 'jotai'
-import type { SetStateAction, WritableAtom } from 'jotai'
-import { atomFamily, useUpdateAtom } from 'jotai/utils'
+import type { SetStateAction, WritableAtom, useSetAtom } from 'jotai'
 import { getTestProvider } from '../testUtils'
 
 const Provider = getTestProvider()
@@ -239,7 +238,7 @@ it('a derived atom from an async atomFamily (#351)', async () => {
   const derivedAtom = atom((get) => get(getAsyncAtom(get(countAtom))))
 
   const Counter = () => {
-    const setCount = useUpdateAtom(countAtom)
+    const setCount = useSetAtom(countAtom)
     const [derived] = useAtom(derivedAtom)
     return (
       <>
